backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.3
coverage-badge==1.1.0
cryptography==42.0.5
data_url==1.0.7
dill==0.3.8
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
mailersend==0.5.6
MarkupSafe==2.1.5
mccabe==0.7.0
openai==1.16.1
orjson==3.9.15
packaging==23.2
passlib==1.7.4
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pylint==3.1.0
pytest==8.0.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
